Add post-imputation to main and refactor to match mbs structure
