+++ cons_results/imputation/impute.py
@@ -1,7 +1,11 @@
 import pandas as pd
 from mbs_results.imputation.ratio_of_means import ratio_of_means
 
-from cons_results.imputation.derive_questions import create_q290, derive_q290
+from cons_results.imputation.post_imputation import (
+  create_q290, 
+  derive_q290,
+  rescale_290_case,
+)
 
 
 def impute(
@@ -47,7 +51,15 @@ def impute(
             filters=filter_df,
         )
     )
-
+    
+    dataframe = rescale_290_case(
+      dataframe,
+      config["period"],
+      config["reference"],
+      config["question_no"],
+      config["target"],
+    )
+    
     dataframe = create_q290(
         dataframe,
         config,
@@ -69,6 +81,7 @@ def impute(
 
     dataframe["period"] = dataframe["period"].dt.strftime("%Y%m").astype("int")
     dataframe = dataframe.reset_index(drop=True)  # remove groupby leftovers
+    print(dataframe["is_backdata"])
     dataframe = dataframe[~dataframe["is_backdata"]]  # remove backdata
     dataframe.drop(columns=["is_backdata"], inplace=True)
