+++ cons_results/imputation/post_imputation.py
@@ -65,3 +65,122 @@ def rescale_290_case(
                 ] = entry[adjusted_response]
 
     return df
+
+def create_q290(
+    df: pd.DataFrame,
+    config: dict,
+    reference: str,
+    period: str,
+    question_no: str,
+    adjustedresponse: str,
+    imputation_flag: str,
+) -> pd.DataFrame:
+    """
+    Parameters
+      ----------
+      df: pd.Dataframe
+        The main construction dataframe following the staging module.
+      config: pd.DataFrame
+        The config as a dictionary
+      reference: str
+        Column name containing reference variable
+      period: str
+        Column name containing period variable
+      question_no: str
+        Column name containing the question number variable
+      adjustedresponse: str
+        Column name containing the adjustedresponse variable
+      imputation_flag: str
+        Column name containing the imputation flag variable
+
+
+    Returns
+    -------
+    df
+      A dataframe with rows added for questioncode 290 where these were missing.
+    """
+
+    missing_290 = df.groupby([period, reference]).filter(
+        lambda x: 290 not in x[question_no].values
+    )[[period, reference]]
+
+    missing_290.drop_duplicates(inplace=True)
+
+    missing_290 = missing_290.assign(
+        question_no=290, adjustedresponse=0.0, imputation_flag="d"
+    )
+
+    missing_290.set_index([period, reference], inplace=True)
+
+    contributors_cols = list(
+        set(config["contributors_keep_cols"] + config["finalsel_keep_cols"])
+    )
+    contributors_df = df[contributors_cols]
+    contributors_df = contributors_df.groupby([period, reference]).first()
+
+    missing_290 = pd.concat(
+        [missing_290, contributors_df.loc[missing_290.index]], axis=1
+    )
+    missing_290.reset_index(inplace=True)
+
+    df = pd.concat([df, missing_290]).reset_index(drop=True)
+
+    return df
+
+
+def derive_q290(
+    df: pd.DataFrame,
+    question_no: str,
+    imputation_flag: str,
+    period: str,
+    reference: str,
+    adjustedresponse: str,
+) -> pd.DataFrame:
+    """
+    Parameters
+      ----------
+      df: pd.Dataframe
+        The main construction dataframe following the staging module.
+      question_no: str
+        Column name containing the question number variable
+      imputation_flag: str
+        Column name containing the imputation flag variable
+      period: str
+        Column name containing the period variable
+      reference: str
+        Column name containing reference variable
+      adjustedresponse: str
+        Column name containing the adjustedresponse variable
+
+
+      Returns
+      -------
+      df
+        A dataframe with 290 derived in rows where components were imputed.
+    """
+
+    imputed_components = (
+        (df[question_no] != 290)
+        & (df[imputation_flag] != "r")
+        & (df["290_flag"] == False)
+    )
+
+    imputed_sums = (
+        df[imputed_components]
+        .groupby([period, reference])[adjustedresponse]
+        .sum()
+        .reset_index()
+        .rename(columns={adjustedresponse: "imputed_sum"})
+    )
+
+    df = df.merge(imputed_sums, on=[period, reference], how="left")
+
+    q290_mask = df[question_no] == 290
+    df["imputed_sum"].fillna(df[adjustedresponse], inplace=True)
+    df.loc[q290_mask, adjustedresponse] = df.loc[q290_mask, "imputed_sum"]
+
+    df.loc[(q290_mask) & (df[imputation_flag] != "r"), imputation_flag] = "d"
+
+    df = df.drop(columns=["imputed_sum"])
+
+    return df
