+++ tests/data/outputs/staged_data.csv
@@ -1,23 +1,15 @@
-,reference,period,questioncode,adjustedresponse,response,frosic2007,froempment,frotover,cell_no,formtype,imputation_flags_adjustedresponse,status,statusencoded,live_adjustedresponse,converted_frotover,imputation_class
-0,10000000000,2022-12-01,211,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,"[1, 7]"
-1,10000000000,2022-12-01,290,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,"[1, 7]"
-2,10000000001,2022-12-01,212,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,"[1, 7]"
-3,10000000001,2022-12-01,290,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,"[1, 7]"
-4,10000000000,2023-01-01,211,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,"[1, 7]"
-5,10000000000,2023-01-01,210,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,"[1, 7]"
-6,10000000000,2023-01-01,290,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,"[1, 7]"
-7,10000000000,2023-02-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-8,10000000000,2023-02-01,210,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-9,10000000000,2023-02-01,290,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-10,10000000000,2023-03-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-11,10000000000,2023-03-01,210,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-12,10000000000,2023-03-01,290,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-13,10000000001,2023-01-01,211,0.0,99999,999,999,999.0,6,9999,,Clear,210.0,0.0,83250.0,"[1, 7]"
-14,10000000001,2023-01-01,210,0.0,99999,999,999,999.0,6,9999,,Clear,210.0,0.0,83250.0,"[1, 7]"
-15,10000000001,2023-01-01,290,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,"[1, 7]"
-16,10000000001,2023-02-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-17,10000000001,2023-02-01,210,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-18,10000000001,2023-02-01,290,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-19,10000000001,2023-03-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-20,10000000001,2023-03-01,210,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
-21,10000000001,2023-03-01,290,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,"[1, 7]"
+,reference,period,questioncode,adjustedresponse,response,frosic2007,froempment,frotover,cell_no,formtype,imputation_flags_adjustedresponse,status,statusencoded,live_adjustedresponse,converted_frotover,imputation_class,290_flag
+0,10000000000,2022-12-01,211,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,0.0,False
+1,10000000000,2022-12-01,290,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,0.0,False
+2,10000000001,2022-12-01,212,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,0.0,False
+3,10000000001,2022-12-01,290,999.0,999,999,999,999.0,6,9999,r,,,999.0,83250.0,0.0,False
+4,10000000000,2023-01-01,211,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,0.0,False
+5,10000000000,2023-01-01,210,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,0.0,False
+6,10000000000,2023-01-01,290,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,0.0,False
+7,10000000000,2023-02-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,0.0,False
+8,10000000000,2023-03-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,0.0,False
+9,10000000001,2023-01-01,211,0.0,99999,999,999,999.0,6,9999,,Clear,210.0,0.0,83250.0,0.0,True
+10,10000000001,2023-01-01,210,0.0,99999,999,999,999.0,6,9999,,Clear,210.0,0.0,83250.0,0.0,True
+11,10000000001,2023-01-01,290,99999.0,99999,999,999,999.0,6,9999,,Clear,210.0,99999.0,83250.0,0.0,True
+12,10000000001,2023-02-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,0.0,False
+13,10000000001,2023-03-01,211,,,999,999,999.0,6,9999,,Clear,210.0,,83250.0,0.0,False
